Use category.slug as a key
This fixes duplicate key error in single category component

The single category component lists all the categories in the sidebar. The used key for the v-for cicle, though, was the entire category object which is wrong. Replacing that with the category.slug value, which should be unique, fixes the error.